chore: move breadcrumbsItemsAtom intp separate file
